# CodeVimarsh-Backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Code Vimarsh API
  version: 1.0.0
  description: Backend API documentation for Code Vimarsh – MSUB Coding Club
servers:
  - url: http://localhost:5000/api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /events:
    get:
      summary: Get all events
      tags:
        - Events
      responses:
        "200":
          description: List of all events
    post:
      summary: Create a new event
      tags:
        - Events
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvent"
      responses:
        "201":
          description: Event created
  /events/{id}:
    get:
      summary: Get an event by ID
      tags:
        - Events
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
        "404":
          description: Event not found
    put:
      summary: Update an event by ID
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEvent"
      responses:
        "200":
          description: Event updated
    delete:
      summary: Delete an event by ID
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event deleted
  /events/{id}/register:
    post:
      summary: Register for an event
      tags:
        - Events
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterEvent"
      responses:
        "201":
          description: Successfully registered
        "404":
          description: Event not found
        "409":
          description: Already registered
...